+++ rpc-client-api/src/custom_error.rs
@@ -100,12 +100,6 @@ pub struct EpochRewardsPeriodActiveErrorData {
     pub slot: u64,
 }
 
-#[derive(Debug, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct SlotNotEpochBoundaryErrorData {
-    pub slot: u64,
-}
-
 impl From<EncodeError> for RpcCustomError {
     fn from(err: EncodeError) -> Self {
         match err {
@@ -253,7 +247,9 @@ impl From<RpcCustomError> for Error {
                     "Rewards cannot be found because slot {slot} is not the epoch boundary. This \
                      may be due to gap in the queried node's local ledger or long-term storage"
                 ),
-                data: Some(serde_json::json!(SlotNotEpochBoundaryErrorData { slot })),
+                data: Some(serde_json::json!({
+                    "slot": slot,
+                })),
             },
             RpcCustomError::LongTermStorageUnreachable => Self {
                 code: ErrorCode::ServerError(JSON_RPC_SERVER_ERROR_LONG_TERM_STORAGE_UNREACHABLE),